MAGN-4858 Adaptive components placed at origin when points are null
The change includes a fix of MAGN-4858.
Root Cause: For creating a adaptive component, multiple RevitAPI are
called in one transaction. The first step is to create a original
component, and then call another RevitAPI to set its position. The issue
happens because an exception throws out in the process of setting
position due to the null point, which finally leads to the uncompleted
creation.
Solution: Using a subtransaction to roll back the creation if an
exception happens.