Fixes Version Selector UI defect MAGN-7250
When user select specific Dynamo version using version selector, we no
longer need the version selector UI in the current session of Revit and
hence it was disabled before. This PR now hides the Ribbon Panel for the
version selector so that we don't show two Dynamo icon on Ribbon.
For Dynamo 0.8.2 onwards, the Dynamo Versions panel is renamed to
"Visual Programming" and the Dynamo button is shown at the same
location. This can't be done for older version of Dynamo, because the
older version of Dynamo always create a new panel named with "Visual
Programming" and if this panel already exists, it throws an error.
0.8.2 onwards DynamoRevitApp checks for any available panel with this
name and tries to add the Dynamo button to existing panel.