Category selection Dropdown - handle spaces
Un-except the exceptional treatment of Category dropdown combos in
DynamoRevit.

The problem was that the ChoiceList contains user-readable strings for
Category names, while the values obtained and sent to DynamoRevit are
expected to be mangled names. Anything with spaces, dashes, etc in its
name is affected by this discrepancy.

This may be just a temp fix, but we force the UpdateValueCore call through
the base class's execution path, which behaves "normally".