Add ElementsByCategory API to View node.
ElementsByCategory is a public API on the View node that given a Category will return the list of elements contained in the view that match that category. An emtpy list is returned if no elements match.   An ArgumentException is thrown if a null or invalid Category Id is passed in.