Add new Select node "All Elements of Category In View" (#2408)
* Add ability to build with Visual Studio 2019

* Add ElementsByCategory API to View node.

ElementsByCategory is a public API on the View node that given a Category will return the list of elements contained in the view that match that category. An emtpy list is returned if no elements match.   An ArgumentException is thrown if a null or invalid Category Id is passed in.

* Back out addition of VS2019 solution, can work with existing in 2019.

* fix typo

* Add new Select node "All Elements of Category In View"

- Takes a View and Category as input
- Returns the list of elements of the specified category in the specified view. Empty list if none found.

In addition, this submission consolidates all sln files to a single one:

- Deleted all but "DynamoRevit.2017.sln.*"
- Renamed "DynamoRevit.2017.sln.*" to "DynamoRevit.sln"
- Updated Readme.md

* Remove unneeded changes

* one more cosmetic change not needed

* Add system test for new node
Rename dynamorevit solution to match Dynamo naming conventions
updated build.xml for 2019

* Update sln name