DYN-6775 Localizing SplashScreen
The buttons from the Static SplashScreen are being localized in DynamoSandbox but in Revit 2026 were not being localized due that the function OnDetectLanguage() was not being called.
Then my fix was just calling this method so the SetLabels methods will be also called and the buttons will be in the expected language.