REVIT-242871 : add node view customization for family instance selection (#3236)
### Purpose
We need this class to simply exist in order to apply default selection
customization that exists in Dynamo Core.
( There is a mechanism in Dynamo Core that relies on the existence of
these NodeViewCustomizations package ).
The bug is not new, it was just exposed now.
Previously this customization class existed in GenerativeDesign.Revit
nodes and that's why it was working in 2026 and earlier versions. E.g in
2026 - if you remove GenerativeDesign.Revit extension
SelectFamilyInstance nodes from D4R will stop working.

### Declarations

Check these if you believe they are true

- [ ] The code base is in a better state after this PR
- [ ] Is documented according to the
[standards](https://github.com/DynamoDS/Dynamo/wiki/Coding-Standards)
- [ ] The level of testing this PR includes is appropriate
- [ ] User facing strings, if any, are extracted into `*.resx` files
- [ ] Snapshot of UI changes, if any.

### FYIs
@QilongTang @saintentropy @johnpierson

Co-authored-by: Bogdan Zavu <bogdan.zavu@autodesk.com>