DYN-9845 : set categories value from id (#3256)
### Purpose
We want to be able to use the category built-in id if possible.
The usage context here is that the AI ( through DynamoMCP ) is aware of
the built-in ids but it is more error prone when it comes to categories
names since those are localized.

### Declarations

Check these if you believe they are true

- [ ] The code base is in a better state after this PR
- [ ] Is documented according to the
[standards](https://github.com/DynamoDS/Dynamo/wiki/Coding-Standards)
- [ ] The level of testing this PR includes is appropriate
- [ ] User facing strings, if any, are extracted into `*.resx` files
- [ ] Snapshot of UI changes, if any.

### Reviewers
@Mikhinja 

### FYIs
@johnpierson

---------

Co-authored-by: Bogdan Zavu <bogdan.zavu@autodesk.com>